update test, get it passing
