+++ mono_repo/test/mono_config_test.dart
@@ -1,0 +1,185 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// Doing a copy-paste from JSON – which uses double-quotes
+// ignore_for_file: prefer_single_quotes
+
+import 'package:mono_repo/src/mono_config.dart';
+import 'package:test/test.dart';
+import 'package:yaml/yaml.dart' as y;
+
+import 'package:mono_repo/src/utils.dart';
+import 'shared.dart';
+
+Matcher throwsArgumentErrorWith(String value) =>
+    throwsA((Object e) => (e as ArgumentError).message == value);
+
+void main() {
+  group('MonoConfig', () {
+    test('sdk version is required', () {
+      expect(
+          () => new MonoConfig.parse('a', {}),
+          throwsArgumentErrorWith(
+              'At least one SDK version is required under "dart".'));
+    });
+
+    test('no stages - end up with one `unit_test` stage with one `test` task',
+        () {
+      var config = new MonoConfig.parse('a', {
+        'dart': ['stable']
+      });
+
+      var oneJob = config.jobs.single;
+      expect(oneJob.sdk, 'stable');
+      expect(oneJob.task.name, 'test');
+      expect(oneJob.task.args, isNull);
+      expect(oneJob.task.config, isNull);
+      expect(oneJob.stageName, 'unit_test');
+    });
+
+    test('valid example', () {
+      var travisYaml = y.loadYaml(testConfig1) as Map<String, dynamic>;
+
+      var config = new MonoConfig.parse('a', travisYaml);
+
+      expect(config.sdks, unorderedEquals(['dev', 'stable', '1.23.0']));
+
+      var jobs = config.jobs.toList();
+
+      expect(encodeJson(jobs), encodeJson(_testConfig1expectedOutput));
+    });
+  });
+}
+
+List get _testConfig1expectedOutput => [
+      {
+        "package": "a",
+        "sdk": "dev",
+        "stageName": "analyze_and_format",
+        "task": {
+          "name": "dartanalyzer",
+          "args": "--fatal-infos --fatal-warnings ."
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "1.23.0",
+        "stageName": "analyze_and_format",
+        "task": {
+          "name": "dartanalyzer",
+          "args": "--fatal-infos --fatal-warnings ."
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "dev",
+        "stageName": "analyze_and_format",
+        "task": {"name": "dartfmt"}
+      },
+      {
+        "package": "a",
+        "sdk": "dev",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--platform chrome",
+          "config": {"xvfb": true}
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "stable",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--platform chrome",
+          "config": {"xvfb": true}
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "1.23.0",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--platform chrome",
+          "config": {"xvfb": true}
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "dev",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--preset travis --total-shards 5 --shard-index 0",
+          "config": {"xvfb": true}
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "stable",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--preset travis --total-shards 5 --shard-index 0",
+          "config": {"xvfb": true}
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "1.23.0",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--preset travis --total-shards 5 --shard-index 0",
+          "config": {"xvfb": true}
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "dev",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--preset travis --total-shards 5 --shard-index 1"
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "stable",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--preset travis --total-shards 5 --shard-index 1"
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "1.23.0",
+        "stageName": "unit_test",
+        "task": {
+          "name": "test",
+          "args": "--preset travis --total-shards 5 --shard-index 1"
+        }
+      },
+      {
+        "package": "a",
+        "sdk": "dev",
+        "stageName": "unit_test",
+        "task": {"name": "test"}
+      },
+      {
+        "package": "a",
+        "sdk": "stable",
+        "stageName": "unit_test",
+        "task": {"name": "test"}
+      },
+      {
+        "package": "a",
+        "sdk": "1.23.0",
+        "stageName": "unit_test",
+        "task": {"name": "test"}
+      }
+    ];
