+++ mono_repo/test/shared.dart
@@ -11,42 +11,27 @@ Matcher throwsUserExceptionWith(String content) => throwsA(
     allOf(isUserException, (e) => (e as UserException).message == content));
 
 final testConfig1 = r'''
-language: dart
 dart:
- - dev
- - stable
- - 1.23.0
+  - dev
+  - stable
+  - 1.23.0
 
-env:
- - FORCE_TEST_EXIT=true
-
-# Content shell needs these fonts.
-addons:
-  something: also here for completeness
-
-before_install:
-  - ignored for now
-  - just here for completeness
-
-dart_task:
- - test: --platform chrome
-   xvfb: true
- - test: --preset travis --total-shards 5 --shard-index 0
-   xvfb: true
- - test: --preset travis --total-shards 5 --shard-index 1
- - test #no args
- - dartanalyzer: --fatal-infos --fatal-warnings .
-
-matrix:
-  exclude:
-    - dart: stable
-      dart_task: dartanalyzer
-  include:
-    - dart: dev
-      dart_task: dartfmt
-  allow_failures:
-    - dart: stable
-      dart_task: dartfmt
+stages:
+  - analyze_and_format:
+    - dartanalyzer: --fatal-infos --fatal-warnings .
+      dart:
+        - dev
+        - 1.23.0
+    - dartfmt:
+      dart:
+        - dev
+  - unit_test:
+    - test: --platform chrome
+      xvfb: true
+    - test: --preset travis --total-shards 5 --shard-index 0
+      xvfb: true
+    - test: --preset travis --total-shards 5 --shard-index 1
+    - test #no args
 ''';
 
 final testConfig2 = r'''
